+++ protocol/requests/create_account.go
@@ -2,6 +2,7 @@ package requests
 
 import (
 	"github.com/ethereum/go-ethereum/common"
+
 	"github.com/pkg/errors"
 
 	utils "github.com/status-im/status-go/common"
