+++ services/wallet/tokenbalances/storage.go
@@ -1,5 +1,7 @@
 package tokenbalances
 
+//go:generate go tool mockgen -package=mock_tokenbalances -source=storage.go -destination=mock/storage/storage.go
+
 import (
 	"context"
 	"math/big"
